reduce duplication rate
