+++ server/src/main/java/org/apache/iotdb/db/mpp/execution/operator/process/RawDataAggregationOperator.java
@@ -20,25 +20,13 @@
 package org.apache.iotdb.db.mpp.execution.operator.process;
 
 import org.apache.iotdb.db.mpp.aggregation.Aggregator;
-import org.apache.iotdb.db.mpp.aggregation.timerangeiterator.ITimeRangeIterator;
 import org.apache.iotdb.db.mpp.execution.operator.Operator;
 import org.apache.iotdb.db.mpp.execution.operator.OperatorContext;
 import org.apache.iotdb.db.mpp.plan.planner.plan.parameter.GroupByTimeParameter;
-import org.apache.iotdb.tsfile.file.metadata.enums.TSDataType;
-import org.apache.iotdb.tsfile.read.common.TimeRange;
-import org.apache.iotdb.tsfile.read.common.block.TsBlock;
-import org.apache.iotdb.tsfile.read.common.block.TsBlockBuilder;
 
-import com.google.common.util.concurrent.ListenableFuture;
-
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
 
-import static org.apache.iotdb.db.mpp.execution.operator.AggregationUtil.appendAggregationResult;
 import static org.apache.iotdb.db.mpp.execution.operator.AggregationUtil.calculateAggregationFromRawData;
-import static org.apache.iotdb.db.mpp.execution.operator.AggregationUtil.initTimeRangeIterator;
 
 /**
  * RawDataAggregationOperator is used to process raw data tsBlock input calculating using value
@@ -50,23 +38,7 @@ import static org.apache.iotdb.db.mpp.execution.operator.AggregationUtil.initTim
  *
  * <p>Return aggregation result in one time interval once.
  */
-public class RawDataAggregationOperator implements ProcessOperator {
-
-  private final OperatorContext operatorContext;
-  private final boolean ascending;
-
-  private final Operator child;
-  private TsBlock inputTsBlock;
-  private boolean canCallNext;
-
-  private final ITimeRangeIterator timeRangeIterator;
-  // current interval of aggregation window [curStartTime, curEndTime)
-  private TimeRange curTimeRange;
-
-  private final List<Aggregator> aggregators;
-
-  // using for building result tsBlock
-  private final TsBlockBuilder resultTsBlockBuilder;
+public class RawDataAggregationOperator extends SingleInputAggregationOperator {
 
   public RawDataAggregationOperator(
       OperatorContext operatorContext,
@@ -74,82 +46,11 @@ public class RawDataAggregationOperator implements ProcessOperator {
       Operator child,
       boolean ascending,
       GroupByTimeParameter groupByTimeParameter) {
-    this.operatorContext = operatorContext;
-    this.ascending = ascending;
-    this.child = child;
-    this.aggregators = aggregators;
-
-    this.timeRangeIterator = initTimeRangeIterator(groupByTimeParameter, ascending, true);
-
-    List<TSDataType> dataTypes = new ArrayList<>();
-    for (Aggregator aggregator : aggregators) {
-      dataTypes.addAll(Arrays.asList(aggregator.getOutputType()));
-    }
-    this.resultTsBlockBuilder = new TsBlockBuilder(dataTypes);
+    super(operatorContext, aggregators, child, ascending, groupByTimeParameter, true);
   }
 
   @Override
-  public OperatorContext getOperatorContext() {
-    return operatorContext;
-  }
-
-  @Override
-  public ListenableFuture<?> isBlocked() {
-    return child.isBlocked();
-  }
-
-  @Override
-  public boolean hasNext() {
-    return curTimeRange != null || timeRangeIterator.hasNextTimeRange();
-  }
-
-  @Override
-  public TsBlock next() {
-    // start stopwatch
-    long maxRuntime = operatorContext.getMaxRunTime().roundTo(TimeUnit.NANOSECONDS);
-    long start = System.nanoTime();
-
-    // reset operator state
-    resultTsBlockBuilder.reset();
-    canCallNext = true;
-
-    while (System.nanoTime() - start < maxRuntime
-        && (curTimeRange != null || timeRangeIterator.hasNextTimeRange())
-        && !resultTsBlockBuilder.isFull()) {
-      if (curTimeRange == null && timeRangeIterator.hasNextTimeRange()) {
-        // move to next time window
-        curTimeRange = timeRangeIterator.nextTimeRange();
-
-        // clear previous aggregation result
-        for (Aggregator aggregator : aggregators) {
-          aggregator.updateTimeRange(curTimeRange);
-        }
-      }
-
-      // calculate aggregation result on current time window
-      if (!calculateNextAggregationResult()) {
-        break;
-      }
-    }
-
-    if (resultTsBlockBuilder.getPositionCount() > 0) {
-      return resultTsBlockBuilder.build();
-    } else {
-      return null;
-    }
-  }
-
-  @Override
-  public boolean isFinished() {
-    return !this.hasNext();
-  }
-
-  @Override
-  public void close() throws Exception {
-    child.close();
-  }
-
-  private boolean calculateNextAggregationResult() {
+  protected boolean calculateNextAggregationResult() {
     while (!calculateAggregationFromRawData(inputTsBlock, aggregators, curTimeRange, ascending)) {
       inputTsBlock = null;
 
@@ -169,10 +70,5 @@ public class RawDataAggregationOperator implements ProcessOperator {
     updateResultTsBlock();
 
     return true;
-  }
-
-  private void updateResultTsBlock() {
-    curTimeRange = null;
-    appendAggregationResult(resultTsBlockBuilder, aggregators, timeRangeIterator);
   }
 }
