+++ server/src/main/java/org/apache/iotdb/db/mpp/execution/operator/source/SeriesScanUtil.java
@@ -29,8 +29,6 @@ import org.apache.iotdb.db.query.reader.universal.PriorityMergeReader;
 import org.apache.iotdb.db.utils.FileLoaderUtils;
 import org.apache.iotdb.db.utils.QueryUtils;
 import org.apache.iotdb.tsfile.exception.write.UnSupportedDataTypeException;
-import org.apache.iotdb.tsfile.file.metadata.AlignedChunkMetadata;
-import org.apache.iotdb.tsfile.file.metadata.AlignedTimeSeriesMetadata;
 import org.apache.iotdb.tsfile.file.metadata.IChunkMetadata;
 import org.apache.iotdb.tsfile.file.metadata.ITimeSeriesMetadata;
 import org.apache.iotdb.tsfile.file.metadata.enums.TSDataType;
@@ -57,6 +55,8 @@ import java.util.PriorityQueue;
 import java.util.Set;
 import java.util.function.ToLongFunction;
 import java.util.stream.Collectors;
+
+import static com.google.common.base.Preconditions.checkArgument;
 
 public class SeriesScanUtil {
   private final FragmentInstanceContext context;
@@ -226,18 +226,13 @@ public class SeriesScanUtil {
     return firstTimeSeriesMetadata.getStatistics();
   }
 
-  Statistics currentFileStatistics(int index) throws IOException {
-    if (!(firstTimeSeriesMetadata instanceof AlignedTimeSeriesMetadata)) {
-      throw new IOException("Can only get statistics by index from alignedTimeSeriesMetaData");
-    }
-    return ((AlignedTimeSeriesMetadata) firstTimeSeriesMetadata).getStatistics(index);
+  protected Statistics currentFileStatistics(int index) throws IOException {
+    checkArgument(index == 0, "Only one sensor in non-aligned SeriesScanUtil.");
+    return currentFileStatistics();
   }
 
-  public Statistics currentFileTimeStatistics() throws IOException {
-    if (!(firstTimeSeriesMetadata instanceof AlignedTimeSeriesMetadata)) {
-      throw new IOException("Can only get statistics of time column from alignedChunkMetaData");
-    }
-    return ((AlignedTimeSeriesMetadata) firstTimeSeriesMetadata).getTimeStatistics();
+  protected Statistics currentFileTimeStatistics() throws IOException {
+    return currentFileStatistics();
   }
 
   boolean currentFileModified() throws IOException {
@@ -362,18 +357,13 @@ public class SeriesScanUtil {
     return firstChunkMetadata.getStatistics();
   }
 
-  Statistics currentChunkStatistics(int index) throws IOException {
-    if (!(firstChunkMetadata instanceof AlignedChunkMetadata)) {
-      throw new IOException("Can only get statistics by index from vectorChunkMetaData");
-    }
-    return ((AlignedChunkMetadata) firstChunkMetadata).getStatistics(index);
+  protected Statistics currentChunkStatistics(int index) throws IOException {
+    checkArgument(index == 0, "Only one sensor in non-aligned SeriesScanUtil.");
+    return currentChunkStatistics();
   }
 
-  Statistics currentChunkTimeStatistics() throws IOException {
-    if (!(firstChunkMetadata instanceof AlignedChunkMetadata)) {
-      throw new IOException("Can only get statistics of time column from alignedChunkMetaData");
-    }
-    return ((AlignedChunkMetadata) firstChunkMetadata).getTimeStatistics();
+  protected Statistics currentChunkTimeStatistics() throws IOException {
+    return currentChunkStatistics();
   }
 
   boolean currentChunkModified() throws IOException {
@@ -590,24 +580,13 @@ public class SeriesScanUtil {
     return firstPageReader.getStatistics();
   }
 
-  Statistics currentPageStatistics(int index) throws IOException {
-    if (firstPageReader == null) {
-      return null;
-    }
-    if (!(firstPageReader.isAlignedPageReader())) {
-      throw new IOException("Can only get statistics by index from AlignedPageReader");
-    }
-    return firstPageReader.getStatistics(index);
+  protected Statistics currentPageStatistics(int index) throws IOException {
+    checkArgument(index == 0, "Only one sensor in non-aligned SeriesScanUtil.");
+    return currentPageStatistics();
   }
 
-  Statistics currentPageTimeStatistics() throws IOException {
-    if (firstPageReader == null) {
-      return null;
-    }
-    if (!(firstPageReader.isAlignedPageReader())) {
-      throw new IOException("Can only get statistics of time column from AlignedPageReader");
-    }
-    return firstPageReader.getTimeStatistics();
+  protected Statistics currentPageTimeStatistics() throws IOException {
+    return currentPageStatistics();
   }
 
   boolean currentPageModified() throws IOException {
@@ -1094,7 +1073,7 @@ public class SeriesScanUtil {
     return orderUtils;
   }
 
-  private class VersionPageReader {
+  protected class VersionPageReader {
 
     protected PriorityMergeReader.MergeReaderPriority version;
     protected IPageReader data;
